# docs/api/permission.openapi.yaml
openapi: 3.1.0
info:
  title: NOVENTI EAOS Permission API
  version: 1.1.17
  description: |
    PHX-K08 Permission Kernel HTTP contract (PHX-G136 adds read-only
    `/permission/roles` catalog listing; PHX-G146/G156/G161 Role→grant product
    posture on `/permission/roles/status`; PHX-G161 env-gated Role→grant live
    mint via `POST /permission/role-grants` when
    `EAOS_ROLE_GRANT_AUTO_WRITE_ENABLED=true` and
    `EAOS_PERMISSION_ROLE_GRANT_MAP` is non-empty; default remains 503
    `GATEWAY_ROLE_GRANT_AUTO_WRITE_DISABLED`).
    tenant_id, subject_id, session_id and platform_scope are derived by
    the trusted authentication boundary and are never accepted from
    request bodies. PHX-G345 adds the explicit, audited, tenant-only
    `/permission/cap-grants` shell. Cap≠grant / title≠permission remain true:
    the supplied capability is an audit label and each request mints a real
    Kernel Grant; it never creates global authority or commercial side effects.
    PHX-G179 documents named Permission status codes (400/403/404/409/503).
    PHX-G185 tightens `RoleGrantProductPosture` to emitted field parity
    (additionalProperties false; milestone const PHX-G161).
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
  /permission/cap-grants:
    post:
      operationId: createCapGrant
      summary: Explicit audited Cap→grant tenant shell
      description: >
        PHX-G345. Grant administrators only. The capability is an audited label,
        not a permission or global role. This endpoint creates one normal
        tenant-scoped Kernel Grant and cannot execute commercial writes.
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CapGrantRequest"
      responses:
        "201":
          description: Tenant grant created or idempotently returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UuidResult"
        "400":
          description: Invalid cap grant payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "401":
          description: Tenant authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "403":
          description: Grant administration authority or tenant eligibility required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "409":
          description: Equivalent grant conflict without an idempotency key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "422":
          description: Closed-schema or non-tenant-scope validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
        default:
          $ref: "#/components/responses/KernelError"
    get:
      operationId: listCapGrants
      summary: List tenant-visible grants for a grant administrator
      parameters:
        - name: principal_subject_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Tenant grants, optionally restricted to one principal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EffectivePermissionList"
        default:
          $ref: "#/components/responses/KernelError"
  /permission/cap-grants/{grantId}/revoke:
    post:
      operationId: revokeCapGrant
      summary: Revoke a real grant through the Cap→grant shell
      parameters:
        - $ref: "#/components/parameters/GrantId"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CapGrantRevokeRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        default:
          $ref: "#/components/responses/KernelError"
...
  /permission/roles/status:
    get:
      operationId: getRoleCatalogStatus
      summary: Read-only Foundation status posture
      description: >
        Roles catalog and grant-map status, including PHX-G146/G156/G161
        role_grant_product posture (auto_grant_from_role_enabled from env;
        auto_write_routes lists /permission/role-grants).
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Roles catalog, grant-map status, and Role→grant product posture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleCatalogStatusEnvelope"
        "401":
          description: Tenant authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        default:
          $ref: "#/components/responses/KernelError"
  /permission/roles:
    get:
      operationId: listPermissionRoles
      summary: Read-only aggregated EAOS roles catalog
      description: >
        Aggregates declared catalog, OIDC map targets, and grant-map keys.
        Never writes grants. Distinct from platform declared-role admin
        (`/platform/roles` in platform.openapi.yaml).
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Aggregated role catalog (flat payload; not a data envelope)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionRoleCatalogResponse"
        "401":
          description: Tenant authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "503":
          description: Role catalog store unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        default:
          $ref: "#/components/responses/KernelError"
  /permission/policies:
    post:
      operationId: createPolicy
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "201":
          description: Policy created in DRAFT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UuidResult"
        "422":
          description: Closed-schema validation failure (extra/unknown fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
        "400":
          description: Invalid policy payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "401":
          description: Tenant authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "409":
          description: Policy conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "503":
          description: Permission service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        default:
          $ref: "#/components/responses/KernelError"
  /permission/policies/{policyId}/activation:
    post:
      operationId: activatePolicy
      parameters:
        - $ref: "#/components/parameters/PolicyId"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VersionedReasonRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "401":
          description: Tenant authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "422":
          description: Closed-schema validation failure (extra/unknown fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
        "400":
          description: Invalid activation payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "404":
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "409":
          description: Version or state conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "503":
          description: Permission service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        default:
          $ref: "#/components/responses/KernelError"
  /permission/policies/{policyId}/deprecation:
    post:
      operationId: deprecatePolicy
      parameters:
        - $ref: "#/components/parameters/PolicyId"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VersionedReasonRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "401":
          description: Tenant authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "422":
          description: Closed-schema validation failure (extra/unknown fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
        "400":
          description: Invalid deprecation payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "404":
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "409":
          description: Version or state conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "503":
          description: Permission service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        default:
          $ref: "#/components/responses/KernelError"
  /permission/grants:
    post:
      operationId: createGrant
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantRequest"
      responses:
        "201":
          description: Direct grant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UuidResult"
        "422":
          description: Closed-schema validation failure (extra/unknown fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
        "400":
          description: Invalid grant payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "401":
          description: Tenant authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "403":
          description: Principal ineligible or cross-tenant forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "409":
          description: Grant conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "503":
          description: Permission service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        default:
          $ref: "#/components/responses/KernelError"
  /permission/grants/{grantId}/revocation:
    post:
      operationId: revokeGrant
      parameters:
        - $ref: "#/components/parameters/GrantId"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VersionedReasonRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "401":
          description: Tenant authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "422":
          description: Closed-schema validation failure (extra/unknown fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
        "400":
          description: Invalid revocation payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "404":
          description: Grant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "409":
          description: Grant already revoked or version conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "503":
          description: Permission service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        default:
          $ref: "#/components/responses/KernelError"
  /permission/grants/{grantId}/delegations:
    post:
      operationId: delegateGrant
      parameters:
        - $ref: "#/components/parameters/GrantId"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DelegateRequest"
      responses:
        "201":
          description: Delegated grant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UuidResult"
        "401":
          description: Tenant authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "422":
          description: Closed-schema validation failure (extra/unknown fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
        "400":
          description: Invalid delegation payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "403":
          description: Delegation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "404":
          description: Grant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "409":
          description: Grant conflict or revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "503":
          description: Permission service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        default:
          $ref: "#/components/responses/KernelError"
  /permission/evaluations:
    post:
      operationId: evaluatePermission
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateRequest"
      responses:
        "200":
          description: Authorization decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateResult"
        "422":
          description: Closed-schema validation failure (extra/unknown fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
        "400":
          description: Invalid evaluate payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "401":
          description: Tenant authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "503":
          description: Permission service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        default:
          $ref: "#/components/responses/KernelError"
  /permission/decisions/{decisionId}/explanation:
    get:
      operationId: explainDecision
      parameters:
        - $ref: "#/components/parameters/DecisionId"
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Decision explanation without secrets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionExplanation"
        "401":
          description: Tenant authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "400":
          description: Invalid explain payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "404":
          description: Decision or grant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "503":
          description: Permission service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        default:
          $ref: "#/components/responses/KernelError"
  /permission/principals/{subjectId}/effective-permissions:
    get:
      operationId: listEffectivePermissions
      description: >
        subjectId is the queried principal, not caller identity. Callers may
        only query themselves unless configured as a trusted evaluation proxy
        or permission auditor.
      parameters:
        - $ref: "#/components/parameters/SubjectId"
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Effective permissions after parent-chain, expiry and scope validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EffectivePermissionList"
        "400":
          description: Invalid list query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "401":
          description: Tenant authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        "503":
          description: Permission service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayDetailError"
        default:
          $ref: "#/components/responses/KernelError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: true
      schema:
        type: string
        minLength: 1
...
  schemas:
...
    UuidResult:
      type: object
      additionalProperties: false
      required: [id, data]
      description: >
        Unified UuidResult dialect (PHX-G170): `id` and `data` carry the same
        UUID for backward-compatible clients of either pre-unification shape.
      properties:
        id:
          type: string
          format: uuid
        data:
          type: string
          format: uuid
        audit_id:
          type: string
          format: uuid
...
    HTTPValidationError:
      type: object
      description: FastAPI request validation envelope (422 detail list).
      additionalProperties: false
      required: [detail]
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationErrorItem"
    GatewayDetailError:
      type: object
      additionalProperties: false
      required: [detail]
      properties:
        detail:
          $ref: "#/components/schemas/ErrorResponse"
...